missing_person_ai_backend/requirements_advanced.txt
# Advanced Face Recognition System Requirements
# Install with: pip install -r requirements_advanced.txt

# Core Face Recognition Libraries
face-recognition==1.3.0
dlib==19.24.2
opencv-python==4.8.1.78

# Machine Learning & Numerical Operations
numpy==1.26.2
scipy==1.11.4

# Image Processing
Pillow==10.1.0

# Deep Learning (Optional - for CNN-based detection)
# Uncomment if you want to use CNN model instead of HOG
# tensorflow==2.15.0
# torch==2.1.1
# torchvision==0.16.1

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0

# JWT Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Utilities
pickle-mixin==1.0.2
